fix: isolate class and support async/defer scripts data
