+++ includes/class-newspack.php
@@ -123,6 +123,7 @@ final class Newspack {
 		include_once NEWSPACK_ABSPATH . 'includes/class-starter-content.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-amp-enhancements.php';
 		include_once NEWSPACK_ABSPATH . 'includes/class-newspack-image-credits.php';
+		include_once NEWSPACK_ABSPATH . 'includes/class-jetpack.php';
 
 		include_once NEWSPACK_ABSPATH . 'includes/class-patches.php';
